Refactoring

+++ NesMenuFolder.cs
@@ -156,10 +156,10 @@ namespace com.clusterrr.hakchi_gui
                 const int maxY = 204;
                 if (image.Width <= maxX && image.Height <= maxY) // Do not upscale
                     return image;
-                if (image.Width / image.Height > maxX / maxY)
-                    outImage = new Bitmap(maxX, maxY * image.Height / image.Width);
+                if ((double)image.Width / (double)image.Height > (double)maxX / (double)maxY)
+                    outImage = new Bitmap(maxX, (int)((double)maxY * (double)image.Height / (double)image.Width));
                 else
-                    outImage = new Bitmap(maxX * image.Width / image.Height, maxY);
+                    outImage = new Bitmap((int)(maxX * (double)image.Width / (double)image.Height), maxY);
                 gr = Graphics.FromImage(outImage);
                 gr.DrawImage(image, new Rectangle(0, 0, outImage.Width, outImage.Height),
                                     new Rectangle(0, 0, image.Width, image.Height), GraphicsUnit.Pixel);
@@ -181,10 +181,10 @@ namespace com.clusterrr.hakchi_gui
                 const int maxY = 40;
                 if (image.Width <= maxX && image.Height <= maxY) // Do not upscale
                     return image;
-                if (image.Width / image.Height > maxX / maxY)
-                    outImage = new Bitmap(maxX, maxY * image.Height / image.Width);
+                if ((double)image.Width / (double)image.Height > (double)maxX / (double)maxY)
+                    outImage = new Bitmap(maxX, (int)((double)maxY * (double)image.Height / (double)image.Width));
                 else
-                    outImage = new Bitmap(maxX * image.Width / image.Height, maxY);
+                    outImage = new Bitmap((int)(maxX * (double)image.Width / (double)image.Height), maxY);
                 gr = Graphics.FromImage(outImage);
                 gr.DrawImage(image, new Rectangle(0, 0, outImage.Width, outImage.Height),
                                     new Rectangle(0, 0, image.Width, image.Height), GraphicsUnit.Pixel);
